地上弱攻撃、地上強攻撃、しゃがみ弱攻撃の実装完了。待機、走り、ジャンプ、しゃがみ、実装済みの攻撃のアニメーションも作成済。動作可能だが一部軽度バグあり。

+++ AnimationComponent.cpp
@@ -71,7 +71,7 @@ void AnimationComponent::Update(float deltaTime)
             if (m_currentClip->loop)
                 m_frameIndex = 0;
             else
-                m_frameIndex = last; // ループなしは最後のフレーム
+                m_frameIndex = (float)last; // ループなしは最後のフレーム
         }
 
         m_sprite->SetFrame(frames[m_frameIndex]);
